renamed files to use camelCase
