+++ packages/cli/commands/config/set/defaultMode.js
